Aula/Aula2.0/Server/requirements.txt
Flask-AppBuilder==4.3.10
Flask-gTTS==0.18
Flask-WTF==1.2.1
